pyproject.toml
[project]
name = "job-post-topic-modelling"
version = "0.0.1"
description = "Research project on categorising job posts"
authors = [{ name = "Adam Hallengreen", email = "adamhallengreen@econ.ku.dk" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "dvc>=3.61.0",
    "dvclive>=3.48.4",
    "hydra-core>=1.3.2",
    "lingua-language-detector>=2.0.2",
    "omegaconf>=2.3.0",
    "pandas>=2.3.1",
    "polars[calamine]>=1.32.2",
    "polars-lts-cpu[calamine]>=1.32.2",
    "tqdm>=4.67.1",
    "bertopic>=0.17.3",
    "hf-xet>=1.1.7",
    "matplotlib>=3.9.4",
    "kaleido>=1.0.0",
    "plotly>=6.3.0",
    "nltk>=3.9.1",
]

[project.urls]
Homepage = "https://AdamHallengreen.github.io/job-post-topic-modelling/"
Repository = "https://github.com/AdamHallengreen/job-post-topic-modelling"
Documentation = "https://AdamHallengreen.github.io/job-post-topic-modelling/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "ty>=0.0.1a16",
    "ruff>=0.11.5",
    "ipykernel>=6.30.1",
    "notebook>=7.4.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_post_topic_modelling"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.9"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.uv]
package = true
